feat(shared): extract configloader and connectivity helpers
Pulls pure file/yaml/koanf helpers from cmd/agentdock/config.go into
shared/configloader, and external-service checks (GitHub, Slack, Redis,
Redis secret beacon) from cmd/agentdock/preflight.go into
shared/connectivity. Neither shared package depends on any agentdock
config struct; types are preserved via parameterization.

This keeps app/config and worker/config able to reuse these helpers
(Phase 4) without duplicating the pure logic.
