+++ cmd/agentdock/prompts.go
@@ -3,18 +3,13 @@ package main
 import (
 	"bufio"
 	"bytes"
-	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"net/http"
 	"os"
 	"os/exec"
 	"strings"
 	"syscall"
-	"time"
 
-	"github.com/redis/go-redis/v9"
 	"golang.org/x/term"
 )
 
@@ -22,77 +17,6 @@ var (
 	stderr  = os.Stderr
 	scanner = bufio.NewScanner(os.Stdin)
 )
-
-// checkRedis verifies connectivity to a Redis server at addr.
-func checkRedis(addr string) error {
-	if addr == "" {
-		return errors.New("address is empty")
-	}
-	client := redis.NewClient(&redis.Options{Addr: addr})
-	defer client.Close()
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	return client.Ping(ctx).Err()
-}
-
-// checkGitHubToken verifies the token can authenticate and has repository access.
-// It returns the authenticated user's login name on success.
-func checkGitHubToken(token string) (string, error) {
-	if token == "" {
-		return "", errors.New("token is empty")
-	}
-
-	httpClient := &http.Client{Timeout: 10 * time.Second}
-	doReq := func(url string) (*http.Response, error) {
-		req, err := http.NewRequest(http.MethodGet, url, nil)
-		if err != nil {
-			return nil, err
-		}
-		req.Header.Set("Authorization", "Bearer "+token)
-		req.Header.Set("Accept", "application/vnd.github+json")
-		return httpClient.Do(req)
-	}
-
-	// Step 1: verify identity.
-	resp, err := doReq("https://api.github.com/user")
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == http.StatusUnauthorized {
-		return "", errors.New("invalid or expired token")
-	}
-	if resp.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("GitHub /user returned HTTP %d", resp.StatusCode)
-	}
-
-	var user struct {
-		Login string `json:"login"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&user); err != nil {
-		return "", fmt.Errorf("decode /user response: %w", err)
-	}
-	login := user.Login
-
-	// Step 2: verify repository access.
-	resp2, err := doReq("https://api.github.com/user/repos?per_page=1")
-	if err != nil {
-		return "", err
-	}
-	defer resp2.Body.Close()
-
-	if resp2.StatusCode == http.StatusForbidden || resp2.StatusCode == http.StatusNotFound {
-		return "", fmt.Errorf("token lacks repository access (user: %s)", login)
-	}
-	if resp2.StatusCode != http.StatusOK {
-		return "", fmt.Errorf("GitHub /user/repos returned HTTP %d", resp2.StatusCode)
-	}
-
-	return login, nil
-}
 
 // checkAgentCLI verifies that the named CLI binary is available and returns
 // the first line of its --version output (stdout+stderr combined).
@@ -154,38 +78,4 @@ func promptHidden(prompt string) string {
 func promptYesNo(prompt string) bool {
 	answer := promptLine(fmt.Sprintf("%s [Y/n]: ", prompt))
 	return answer == "" || strings.ToLower(answer) == "y" || strings.ToLower(answer) == "yes"
-}
-
-// checkSlackToken verifies the bot token via Slack auth.test API.
-// Returns the bot's user_id on success.
-func checkSlackToken(token string) (string, error) {
-	if token == "" {
-		return "", errors.New("token is empty")
-	}
-	if !strings.HasPrefix(token, "xoxb-") {
-		return "", errors.New("Slack bot token must start with xoxb-")
-	}
-
-	httpClient := &http.Client{Timeout: 10 * time.Second}
-	req, _ := http.NewRequest(http.MethodPost, "https://slack.com/api/auth.test", nil)
-	req.Header.Set("Authorization", "Bearer "+token)
-
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	var body struct {
-		OK     bool   `json:"ok"`
-		UserID string `json:"user_id"`
-		Error  string `json:"error"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&body); err != nil {
-		return "", err
-	}
-	if !body.OK {
-		return "", fmt.Errorf("auth.test failed: %s", body.Error)
-	}
-	return body.UserID, nil
 }
