+++ cmd/agentdock/prompts_test.go
@@ -1,15 +1,19 @@
 package main
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/Ivantseng123/agentdock/shared/connectivity"
+)
 
 func TestCheckSlackToken_RejectsEmpty(t *testing.T) {
-	if _, err := checkSlackToken(""); err == nil {
+	if _, err := connectivity.CheckSlackToken(""); err == nil {
 		t.Error("expected error for empty token")
 	}
 }
 
 func TestCheckSlackToken_RejectsBadPrefix(t *testing.T) {
-	if _, err := checkSlackToken("not-a-slack-token"); err == nil {
+	if _, err := connectivity.CheckSlackToken("not-a-slack-token"); err == nil {
 		t.Error("expected error for token without xoxb- prefix")
 	}
 }
