+++ shared/configloader/path_test.go
@@ -1,0 +1,26 @@
+package configloader
+
+import (
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+)
+
+func TestResolveConfigPath_ExpandsTilde(t *testing.T) {
+	home, err := os.UserHomeDir()
+	if err != nil {
+		t.Skip("no home dir")
+	}
+	got, err := ResolveConfigPath("~/foo.yaml")
+	if err != nil {
+		t.Fatal(err)
+	}
+	want := filepath.Join(home, "foo.yaml")
+	if got != want {
+		t.Errorf("got %q want %q", got, want)
+	}
+	if !strings.HasPrefix(got, home) {
+		t.Errorf("should start with home: %q", got)
+	}
+}
